refactor unit test. each package is a stand alone along with its unit tests
